decked-out-server/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
google_api_python_client==2.70.0
google_auth_oauthlib==0.8.0
openai==0.25.0
protobuf==4.21.12
python-dotenv==0.21.0
gunicorn==20.0.4